pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ego-exo-datakit"
version = "0.1.0"
description = "Small Python reader for ego-exo MCAP files generated by the ego-exo-cam-v2 pipeline."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcap>=1.3.0",
  "mcap-protobuf-support>=0.5.1",
  "numpy>=1.24",
  "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
